Implement workaround to get printf("string without args") working
GCC is replacing printf("string with no args"); with puts("string with no args"); even with no optimizations enabled!

This is apparently not implemented in a usable way for us in Arduino land and results in a) no output or b) an infinitely rebooting application, depending on whether it's my lucky day or not.

Current workaround is to redefine puts() for our needs. Another option is noted in the comment along with the function, in case that bites us in the end.